extract functions to utils.js codes so that it can be tested with jest
only taskboard's formatLocalISOFromString is extracted... for now.
