+++ manager/templates/manager/taskboard.html
@@ -5,7 +5,7 @@
 <link rel="stylesheet" href="{% static 'manager/taskboard.css' %}">
 <link rel="stylesheet" href="{% static 'manager/calendar.css' %}">
 <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js" defer></script>
-<script src="{% static 'manager/taskboard.js' %}" defer></script>
+<script src="{% static 'manager/taskboard.js' %}" defer type="module"></script>
 {% endblock %}
 
 {% block content %}
